proxy.js file renamed to proxy_configuration.js, added code examples, Apify UI input configuration option, validation changes
